requirements.txt
zoom-meeting-sdk
python-dotenv
openai
pyjwt